+++ src/CMakeLists.txt
@@ -5,11 +5,6 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -Wextra -std=c++11")
 set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS} -g -O0")
 set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS} -O3")
 
-set(SRC_FILES
-    evd.cpp
-    svd.cpp
-    util.cpp
-    main.cpp)
-
-add_executable(main ${SRC_FILES})
-target_link_libraries(main m)
+add_subdirectory(utils)
+add_subdirectory(svd)
+add_subdirectory(evd)
