+++ src/evd/classic/main.cpp
@@ -1,0 +1,1 @@
+int main(void) { return 0; }
