+++ src/svd/one-sided/CMakeLists.txt
@@ -1,0 +1,11 @@
+set(SOURCES
+    svd.cpp
+    svd.hpp
+    )
+
+add_library(svd-lib STATIC ${SOURCES})
+target_include_directories(svd-lib PUBLIC .)
+target_link_libraries(svd-lib utils-lib)
+
+add_executable(svd main.cpp)
+target_link_libraries(svd utils-lib m)
