+++ test/evd/classic/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_executable(evd_test evd_test.cpp)
+target_link_libraries(evd_test gtest_main evd-lib utils-lib m)
+
+add_test(NAME EVD COMMAND evd_test)
